Add CsvGrid class
CsvGrid contains functionality related to the grid that displays the contents of the OutputCSV object.
